+++ Gopkg.lock
@@ -1513,6 +1513,6 @@
 [solve-meta]
   analyzer-name = "dep"
   analyzer-version = 1
-  inputs-digest = "4d7140ff09e65a21b5bcfb52300d27fbb7d044fc7fbd2d0d0a1c1c7e3dab7daf"
+  inputs-digest = "69d4f0b300945d3929b747bf5fde5fb8ed31ddcf9343b70d485530383971d706"
   solver-name = "gps-cdcl"
   solver-version = 1
