+++ istio.io/fortio/.circleci/config.yml
@@ -2,21 +2,48 @@
 #
 # Check https://circleci.com/docs/2.0/language-go/ for more details
 version: 2
-jobs:
-  build:
+
+# common setup and steps
+defaultEnv: &defaultEnv
     docker:
       # specify the version
       - image: istio/fortio.build:v6
     working_directory: /go/src/istio.io/fortio
+
+jobs:
+  unit-tests:
+    <<: *defaultEnv
+    steps:
+      - checkout
+      - run: make submodule
+      - run: make test
+  release-tests:
+    <<: *defaultEnv
     steps:
       - setup_remote_docker:
-          version: 17.10.0-ce
+          version: 17.11.0-ce
       - checkout
-      - run: make submodule # will run: (only if needed:)
-#      - run: git submodule sync
-#      - run: git submodule update --init
-      - run: make test
+      - run: make submodule
+      - run: make release-test
+  linters:
+    <<: *defaultEnv
+    steps:
+      - checkout
+      - run: make submodule
       # If ran with default we get random errors because OOM killer kills some linters
       - run: make install local-lint  DEBUG_LINTERS="--concurrency=2"
-      - run: make webtest
+  codecov:
+    <<: *defaultEnv
+    steps:
+      - checkout
+      - run: make submodule
       - run: make coverage
+
+workflows:
+  version: 2
+  all:
+    jobs:
+      - unit-tests
+      - release-tests
+      - codecov
+      - linters
