+++ istio.io/fortio/fhttp/http_client.go
@@ -21,24 +21,15 @@ import (
 	"fmt"
 	"io"
 	"io/ioutil"
-	"math/rand"
 	"net"
 	"net/http"
 	"net/http/httputil"
 	"net/url"
-	"os"
-	"sort"
-	"strconv"
 	"strings"
-	"sync/atomic"
 	"time"
-	"unicode/utf8"
-	// get /debug/pprof endpoints on default server - make sure report only doesn't have it
-	_ "net/http/pprof"
 
 	"istio.io/fortio/fnet"
 	"istio.io/fortio/log"
-	"istio.io/fortio/stats"
 	"istio.io/fortio/version"
 )
 
@@ -53,7 +44,7 @@ type Fetcher interface {
 }
 
 var (
-	// BufferSizeKb size of the buffer (max data) for optimized client in kilobytes defaults to 32k.
+	// BufferSizeKb size of the buffer (max data) for optimized client in kilobytes defaults to 128k.
 	BufferSizeKb = 128
 	// CheckConnectionClosedHeader indicates whether to check for server side connection closed headers.
 	CheckConnectionClosedHeader = false
@@ -61,8 +52,6 @@ var (
 	contentLengthHeader   = []byte("\r\ncontent-length:")
 	connectionCloseHeader = []byte("\r\nconnection: close")
 	chunkedHeader         = []byte("\r\nTransfer-Encoding: chunked")
-	// Start time of the server (used in debug handler for uptime).
-	startTime time.Time
 )
 
 // NewHTTPOptions creates and initialize a HTTPOptions object.
@@ -74,15 +63,12 @@ func NewHTTPOptions(url string) *HTTPOptions {
 }
 
 // Init initializes the headers in an HTTPOptions (User-Agent).
-// It replaces plain % to %25 in the url. If you already have properly
-// escaped URLs use o.URL = to set it.
 func (h *HTTPOptions) Init(url string) *HTTPOptions {
 	if h.initDone {
 		return h
 	}
 	h.initDone = true
-	// unescape then rescape % to %25 (so if it was already %25 it stays)
-	h.URL = strings.Replace(strings.Replace(url, "%25", "%", -1), "%", "%25", -1)
+	h.URL = url
 	h.NumConnections = 1
 	if h.HTTPReqTimeOut <= 0 {
 		h.HTTPReqTimeOut = HTTPReqTimeOutDefaultValue
@@ -139,6 +125,7 @@ type HTTPOptions struct {
 	DisableKeepAlive  bool // so default is keep alive
 	AllowHalfClose    bool // if not keepalive, whether to half close after request
 	Insecure          bool // do not verify certs for https
+	FollowRedirects   bool // For the Std Client only: follow redirects.
 	initDone          bool
 	https             bool // whether URLSchemeCheck determined this was an https:// call or not
 	// ExtraHeaders to be added to each request.
@@ -325,14 +312,36 @@ func NewStdClient(o *HTTPOptions) *Client {
 		&http.Client{
 			Timeout:   o.HTTPReqTimeOut,
 			Transport: &tr,
-			// Lets us see the raw response instead of auto following redirects.
-			CheckRedirect: func(req *http.Request, via []*http.Request) error {
-				return http.ErrUseLastResponse
-			},
 		},
 		&tr,
 	}
+	if !o.FollowRedirects {
+		// Lets us see the raw response instead of auto following redirects.
+		client.client.CheckRedirect = func(req *http.Request, via []*http.Request) error {
+			return http.ErrUseLastResponse
+		}
+	}
 	return &client
+}
+
+// FetchURL fetches the data at the given url using the standard client and default options.
+// Returns the http status code (http.StatusOK == 200 for success) and the data.
+// To be used only for single fetches or when performance doesn't matter as the client is closed at the end.
+func FetchURL(url string) (int, []byte) {
+	o := NewHTTPOptions(url)
+	// Maximize chances of getting the data back, vs the raw payload like the fast client
+	o.DisableFastClient = true
+	o.FollowRedirects = true
+	return Fetch(o)
+}
+
+// Fetch creates a client an performs a fetch according to the http options passed in.
+// To be used only for single fetches or when performance doesn't matter as the client is closed at the end.
+func Fetch(httpOptions *HTTPOptions) (int, []byte) {
+	cli := NewClient(httpOptions)
+	code, data, _ := cli.Fetch()
+	cli.Close()
+	return code, data
 }
 
 // FastClient is a fast, lockfree single purpose http 1.0/1.1 client.
@@ -395,21 +404,12 @@ func NewFastClient(o *HTTPOptions) Fetcher {
 		bc.port = url.Scheme // ie http which turns into 80 later
 		log.LogVf("No port specified, using %s", bc.port)
 	}
-	addrs, err := net.LookupIP(bc.hostname)
-	if err != nil {
-		log.Errf("Unable to lookup '%s' : %v", bc.host, err)
+	addr := fnet.Resolve(bc.hostname, bc.port)
+	if addr == nil {
+		// Error already logged
 		return nil
 	}
-	if len(addrs) > 1 && log.LogDebug() {
-		log.Debugf("Using only the first of the addresses for %s : %v", bc.host, addrs)
-	}
-	log.Debugf("Will go to %s", addrs[0])
-	bc.dest.IP = addrs[0]
-	bc.dest.Port, err = net.LookupPort("tcp", bc.port)
-	if err != nil {
-		log.Errf("Unable to resolve port '%s' : %v", bc.port, err)
-		return nil
-	}
+	bc.dest = *addr
 	// Create the bytes for the request:
 	host := bc.host
 	if o.hostOverride != "" {
@@ -439,155 +439,6 @@ func NewFastClient(o *HTTPOptions) Fetcher {
 	bc.req = buf.Bytes()
 	log.Debugf("Created client:\n%+v\n%s", bc.dest, bc.req)
 	return &bc
-}
-
-// Used for the fast case insensitive search
-const toUpperMask = ^byte('a' - 'A')
-
-// Slow but correct version
-func toUpper(b byte) byte {
-	if b >= 'a' && b <= 'z' {
-		b -= ('a' - 'A')
-	}
-	return b
-}
-
-// ASCIIToUpper returns a byte array equal to the input string but in lowercase.
-// Only wotks for ASCII, not meant for unicode.
-func ASCIIToUpper(str string) []byte {
-	numChars := utf8.RuneCountInString(str)
-	if numChars != len(str) && log.LogVerbose() {
-		log.Errf("ASCIIFold(\"%s\") contains %d characters, some non ascii (byte length %d): will mangle", str, numChars, len(str))
-	}
-	res := make([]byte, numChars)
-	// less surprising if we only mangle the extended characters
-	i := 0
-	for _, c := range str { // Attention: _ here != i for unicode characters
-		res[i] = toUpper(byte(c))
-		i++
-	}
-	return res
-}
-
-// FoldFind searches the bytes assuming ascii, ignoring the lowercase bit
-// for testing. Not intended to work with unicode, meant for http headers
-// and to be fast (see benchmark in test file).
-func FoldFind(haystack []byte, needle []byte) (bool, int) {
-	idx := 0
-	found := false
-	hackstackLen := len(haystack)
-	needleLen := len(needle)
-	if needleLen == 0 {
-		return true, 0
-	}
-	if needleLen > hackstackLen { // those 2 ifs also handles haystackLen == 0
-		return false, -1
-	}
-	needleOffset := 0
-	for {
-		h := haystack[idx]
-		n := needle[needleOffset]
-		// This line is quite performance sensitive. calling toUpper() for instance
-		// is a 30% hit, even if called only on the haystack. The XOR lets us be
-		// true for equality and the & with mask also true if the only difference
-		// between the 2 is the case bit.
-		xor := h ^ n // == 0 if strictly equal
-		if (xor&toUpperMask) != 0 || (((h < 32) || (n < 32)) && (xor != 0)) {
-			idx -= (needleOffset - 1) // does ++ most of the time
-			needleOffset = 0
-			if idx >= hackstackLen {
-				break
-			}
-			continue
-		}
-		if needleOffset == needleLen-1 {
-			found = true
-			break
-		}
-		needleOffset++
-		idx++
-		if idx >= hackstackLen {
-			break
-		}
-	}
-	if !found {
-		return false, -1
-	}
-	return true, idx - needleOffset
-}
-
-// ParseDecimal extracts the first positive integer number from the input.
-// spaces are ignored.
-// any character that isn't a digit cause the parsing to stop
-func ParseDecimal(inp []byte) int {
-	res := -1
-	for _, b := range inp {
-		if b == ' ' && res == -1 {
-			continue
-		}
-		if b < '0' || b > '9' {
-			break
-		}
-		digit := int(b - '0')
-		if res == -1 {
-			res = digit
-		} else {
-			res = 10*res + digit
-		}
-	}
-	return res
-}
-
-// ParseChunkSize extracts the chunk size and consumes the line.
-// Returns the offset of the data and the size of the chunk,
-// 0, -1 when not found.
-func ParseChunkSize(inp []byte) (int, int) {
-	if log.LogDebug() {
-		log.Debugf("ParseChunkSize(%s)", DebugSummary(inp, 128))
-	}
-	res := -1
-	off := 0
-	end := len(inp)
-	inDigits := true
-	for {
-		if off >= end {
-			return off, -1
-		}
-		if inDigits {
-			b := toUpper(inp[off])
-			var digit int
-			if b >= 'A' && b <= 'F' {
-				digit = 10 + int(b-'A')
-			} else if b >= '0' && b <= '9' {
-				digit = int(b - '0')
-			} else {
-				inDigits = false
-				if res == -1 {
-					log.Errf("Didn't find hex number %q", inp)
-					return off, res
-				}
-				continue
-			}
-			if res == -1 {
-				res = digit
-			} else {
-				res = 16*res + digit
-			}
-		} else {
-			// After digits, skipping ahead to find \r\n
-			if inp[off] == '\r' {
-				off++
-				if off >= end {
-					return off, -1
-				}
-				if inp[off] == '\n' {
-					// good case
-					return off + 1, res
-				}
-			}
-		}
-		off++
-	}
 }
 
 // return the result from the state.
@@ -674,24 +525,6 @@ func (c *FastClient) Fetch() (int, []byte, int) {
 	}
 	// Return the result:
 	return c.returnRes()
-}
-
-// EscapeBytes returns printable string. Same as %q format without the
-// surrounding/extra "".
-func EscapeBytes(buf []byte) string {
-	e := fmt.Sprintf("%q", buf)
-	return e[1 : len(e)-1]
-}
-
-// DebugSummary returns a string with the size and escaped first max/2 and
-// last max/2 bytes of a buffer (or the whole escaped buffer if small enough).
-func DebugSummary(buf []byte, max int) string {
-	l := len(buf)
-	if l <= max+3 { //no point in shortening to add ... if we could return those 3
-		return EscapeBytes(buf)
-	}
-	max /= 2
-	return fmt.Sprintf("%d: %s...%s", l, EscapeBytes(buf[:max]), EscapeBytes(buf[l-max:]))
 }
 
 // Response reading:
@@ -881,493 +714,4 @@ func (c *FastClient) readResponse(conn *net.TCPConn, reusedSocket bool) {
 		}
 		// we cleared c.socket in caller already
 	}
-}
-
-// -- Echo Server --
-
-var (
-	// EchoRequests is the number of request received. Only updated in Debug mode.
-	EchoRequests int64
-)
-
-func removeTrailingPercent(s string) string {
-	if strings.HasSuffix(s, "%") {
-		return s[:len(s)-1]
-	}
-	return s
-}
-
-// generateStatus from string, format: status="503" for 100% 503s
-// status="503:20,404:10,403:0.5" for 20% 503s, 10% 404s, 0.5% 403s 69.5% 200s
-func generateStatus(status string) int {
-	lst := strings.Split(status, ",")
-	log.Debugf("Parsing status %s -> %v", status, lst)
-	// Simple non probabilistic status case:
-	if len(lst) == 1 && !strings.ContainsRune(status, ':') {
-		s, err := strconv.Atoi(status)
-		if err != nil {
-			log.Warnf("Bad input status %v, not a number nor comma and colon separated %% list", status)
-			return http.StatusBadRequest
-		}
-		log.Debugf("Parsed status %s -> %d", status, s)
-		return s
-	}
-	weights := make([]float32, len(lst))
-	codes := make([]int, len(lst))
-	lastPercent := float64(0)
-	i := 0
-	for _, entry := range lst {
-		l2 := strings.Split(entry, ":")
-		if len(l2) != 2 {
-			log.Warnf("Should have exactly 1 : in status list %s -> %v", status, entry)
-			return http.StatusBadRequest
-		}
-		s, err := strconv.Atoi(l2[0])
-		if err != nil {
-			log.Warnf("Bad input status %v -> %v, not a number before colon", status, l2[0])
-			return http.StatusBadRequest
-		}
-		percStr := removeTrailingPercent(l2[1])
-		p, err := strconv.ParseFloat(percStr, 32)
-		if err != nil || p < 0 || p > 100 {
-			log.Warnf("Percentage is not a [0. - 100.] number in %v -> %v : %v %f", status, percStr, err, p)
-			return http.StatusBadRequest
-		}
-		lastPercent += p
-		// Round() needed to cover 'exactly' 100% and not more or less because of rounding errors
-		p32 := float32(stats.Round(lastPercent))
-		if p32 > 100. {
-			log.Warnf("Sum of percentage is greater than 100 in %v %f %f %f", status, lastPercent, p, p32)
-			return http.StatusBadRequest
-		}
-		weights[i] = p32
-		codes[i] = s
-		i++
-	}
-	res := 100. * rand.Float32()
-	for i, v := range weights {
-		if res <= v {
-			log.Debugf("[0.-100.[ for %s roll %f got #%d -> %d", status, res, i, codes[i])
-			return codes[i]
-		}
-	}
-	log.Debugf("[0.-100.[ for %s roll %f no hit, defaulting to OK", status, res)
-	return http.StatusOK // default/reminder of probability table
-}
-
-// MaxPayloadSize is the maximum size of payload to be generated by the
-// EchoHandler size= argument. In bytes.
-const MaxPayloadSize = 256 * 1024 // 256k
-
-// generateSize from string, format: "size=512" for 100% 512 bytes body replies,
-// size="512:20,16384:10" for 20% 512 bytes, 10% 16k, 70% default echo back.
-// returns -1 for the default case, so one can specify 0 and force no payload
-// even if it's a post request with a payload (to test asymetric large inbound
-// small outbound).
-// TODO: refactor similarities with status and delay
-func generateSize(sizeInput string) (size int) {
-	size = -1 // default value/behavior
-	if len(sizeInput) == 0 {
-		return size
-	}
-	lst := strings.Split(sizeInput, ",")
-	log.Debugf("Parsing size %s -> %v", sizeInput, lst)
-	// Simple non probabilistic status case:
-	if len(lst) == 1 && !strings.ContainsRune(sizeInput, ':') {
-		s, err := strconv.Atoi(sizeInput)
-		if err != nil {
-			log.Warnf("Bad input size %v, not a number nor comma and colon separated %% list", sizeInput)
-			return size
-		}
-		size = s
-		log.Debugf("Parsed size %s -> %d", sizeInput, size)
-		if size > MaxPayloadSize {
-			log.Warnf("Requested size %d greater than max size %d, using max instead", size, MaxPayloadSize)
-			size = MaxPayloadSize
-		}
-		return size
-	}
-	weights := make([]float32, len(lst))
-	sizes := make([]int, len(lst))
-	lastPercent := float64(0)
-	i := 0
-	for _, entry := range lst {
-		l2 := strings.Split(entry, ":")
-		if len(l2) != 2 {
-			log.Warnf("Should have exactly 1 : in size list %s -> %v", sizeInput, entry)
-			return size
-		}
-		s, err := strconv.Atoi(l2[0])
-		if err != nil {
-			log.Warnf("Bad input size %v -> %v, not a number before colon", sizeInput, l2[0])
-			return size
-		}
-		if s > MaxPayloadSize {
-			log.Warnf("Requested size %d greater than max size %d, using max instead", s, MaxPayloadSize)
-			s = MaxPayloadSize
-		}
-		percStr := removeTrailingPercent(l2[1])
-		p, err := strconv.ParseFloat(percStr, 32)
-		if err != nil || p < 0 || p > 100 {
-			log.Warnf("Percentage is not a [0. - 100.] number in %v -> %v : %v %f", sizeInput, percStr, err, p)
-			return size
-		}
-		lastPercent += p
-		// Round() needed to cover 'exactly' 100% and not more or less because of rounding errors
-		p32 := float32(stats.Round(lastPercent))
-		if p32 > 100. {
-			log.Warnf("Sum of percentage is greater than 100 in %v %f %f %f", sizeInput, lastPercent, p, p32)
-			return size
-		}
-		weights[i] = p32
-		sizes[i] = s
-		i++
-	}
-	res := 100. * rand.Float32()
-	for i, v := range weights {
-		if res <= v {
-			log.Debugf("[0.-100.[ for %s roll %f got #%d -> %d", sizeInput, res, i, sizes[i])
-			return sizes[i]
-		}
-	}
-	log.Debugf("[0.-100.[ for %s roll %f no hit, defaulting to -1", sizeInput, res)
-	return size // default/reminder of probability table
-}
-
-// MaxDelay is the maximum delay allowed for the echoserver responses.
-// 1.5s so we can test the default 1s timeout in envoy.
-const MaxDelay = 1500 * time.Millisecond
-
-// generateDelay from string, format: delay="100ms" for 100% 100ms delay
-// delay="10ms:20,20ms:10,1s:0.5" for 20% 10ms, 10% 20ms, 0.5% 1s and 69.5% 0
-// TODO: very similar with generateStatus - refactor?
-func generateDelay(delay string) time.Duration {
-	lst := strings.Split(delay, ",")
-	log.Debugf("Parsing delay %s -> %v", delay, lst)
-	if len(delay) == 0 {
-		return -1
-	}
-	// Simple non probabilistic status case:
-	if len(lst) == 1 && !strings.ContainsRune(delay, ':') {
-		d, err := time.ParseDuration(delay)
-		if err != nil {
-			log.Warnf("Bad input delay %v, not a duration nor comma and colon separated %% list", delay)
-			return -1
-		}
-		log.Debugf("Parsed delay %s -> %d", delay, d)
-		if d > MaxDelay {
-			d = MaxDelay
-		}
-		return d
-	}
-	weights := make([]float32, len(lst))
-	delays := make([]time.Duration, len(lst))
-	lastPercent := float64(0)
-	i := 0
-	for _, entry := range lst {
-		l2 := strings.Split(entry, ":")
-		if len(l2) != 2 {
-			log.Warnf("Should have exactly 1 : in delay list %s -> %v", delay, entry)
-			return -1
-		}
-		d, err := time.ParseDuration(l2[0])
-		if err != nil {
-			log.Warnf("Bad input delay %v -> %v, not a number before colon", delay, l2[0])
-			return -1
-		}
-		if d > MaxDelay {
-			d = MaxDelay
-		}
-		percStr := removeTrailingPercent(l2[1])
-		p, err := strconv.ParseFloat(percStr, 32)
-		if err != nil || p < 0 || p > 100 {
-			log.Warnf("Percentage is not a [0. - 100.] number in %v -> %v : %v %f", delay, percStr, err, p)
-			return -1
-		}
-		lastPercent += p
-		// Round() needed to cover 'exactly' 100% and not more or less because of rounding errors
-		p32 := float32(stats.Round(lastPercent))
-		if p32 > 100. {
-			log.Warnf("Sum of percentage is greater than 100 in %v %f %f %f", delay, lastPercent, p, p32)
-			return -1
-		}
-		weights[i] = p32
-		delays[i] = d
-		i++
-	}
-	res := 100. * rand.Float32()
-	for i, v := range weights {
-		if res <= v {
-			log.Debugf("[0.-100.[ for %s roll %f got #%d -> %d", delay, res, i, delays[i])
-			return delays[i]
-		}
-	}
-	log.Debugf("[0.-100.[ for %s roll %f no hit, defaulting to 0", delay, res)
-	return 0
-}
-
-// EchoHandler is an http server handler echoing back the input.
-func EchoHandler(w http.ResponseWriter, r *http.Request) {
-	log.LogVf("%v %v %v %v", r.Method, r.URL, r.Proto, r.RemoteAddr)
-	data, err := ioutil.ReadAll(r.Body) // must be done before calling FormValue
-	if err != nil {
-		log.Errf("Error reading %v", err)
-		http.Error(w, err.Error(), http.StatusInternalServerError)
-		return
-	}
-	log.Debugf("Read %d", len(data))
-	dur := generateDelay(r.FormValue("delay"))
-	if dur > 0 {
-		log.LogVf("Sleeping for %v", dur)
-		time.Sleep(dur)
-	}
-	statusStr := r.FormValue("status")
-	var status int
-	if statusStr != "" {
-		status = generateStatus(statusStr)
-	} else {
-		status = http.StatusOK
-	}
-	if log.LogDebug() {
-		// TODO: this easily lead to contention - use 'thread local'
-		rqNum := atomic.AddInt64(&EchoRequests, 1)
-		log.Debugf("Request # %v", rqNum)
-		for name, headers := range r.Header {
-			for _, h := range headers {
-				log.Debugf("Header %v: %v\n", name, h)
-			}
-		}
-	}
-	if r.FormValue("close") != "" {
-		log.Debugf("Adding Connection:close / will close socket")
-		w.Header().Set("Connection", "close")
-	}
-	size := generateSize(r.FormValue("size"))
-	if size >= 0 {
-		log.LogVf("Writing %d size with %d status", size, status)
-		writePayload(w, status, size)
-		return
-	}
-	// echo back the Content-Type and Content-Length in the response
-	for _, k := range []string{"Content-Type", "Content-Length"} {
-		if v := r.Header.Get(k); v != "" {
-			w.Header().Set(k, v)
-		}
-	}
-	w.WriteHeader(status)
-	if _, err = w.Write(data); err != nil {
-		log.Errf("Error writing response %v to %v", err, r.RemoteAddr)
-	}
-}
-
-var payload = make([]byte, MaxPayloadSize)
-
-func init() {
-	// One shared and 'constant' (over time) but pseudo random content for payload
-	// (to defeat compression). We don't need crypto strength here, just low cpu
-	// and speed:
-	rand.Read(payload) // nolint:gas
-}
-
-func writePayload(w http.ResponseWriter, status int, size int) {
-	w.Header().Set("Content-Type", "application/octet-stream")
-	w.Header().Set("Content-Length", strconv.Itoa(size))
-	w.WriteHeader(status)
-	n, err := w.Write(payload[:size])
-	if err != nil || n != size {
-		log.Errf("Error writing payload of size %d: %d %v", size, n, err)
-	}
-}
-
-func closingServer(listener net.Listener) error {
-	var err error
-	for {
-		var c net.Conn
-		c, err = listener.Accept()
-		if err != nil {
-			log.Errf("Accept error in dummy server %v", err)
-			break
-		}
-		log.LogVf("Got connection from %v, closing", c.RemoteAddr())
-		err = c.Close()
-		if err != nil {
-			log.Errf("Close error in dummy server %v", err)
-			break
-		}
-	}
-	return err
-}
-
-// DynamicHTTPServer listens on an available port, sets up an http or https
-// (when secure is true) server on it and returns the listening port and
-// mux to which one can attach handlers to.
-// TODO: merge/refactor/share with Serve()
-func DynamicHTTPServer(secure bool) (int, *http.ServeMux) {
-	m := http.NewServeMux()
-	s := &http.Server{
-		Handler: m,
-	}
-	listener, err := net.Listen("tcp", ":0") // nolint: gas
-	if err != nil {
-		log.Fatalf("Unable to listen to dynamic port: %v", err)
-	}
-	port := listener.Addr().(*net.TCPAddr).Port
-	log.Infof("Using port: %d", port)
-	go func() {
-		var err error
-		if secure {
-			log.Errf("Secure setup not yet supported. Will just close incoming connections for now")
-			//err = http.ServeTLS(listener, nil, "", "") // go 1.9
-			err = closingServer(listener)
-		} else {
-			err = s.Serve(listener)
-		}
-		if err != nil {
-			log.Fatalf("Unable to serve with secure=%v on %d: %v", secure, port, err)
-		}
-	}()
-	return port, m
-}
-
-/*
-// DebugHandlerTemplate returns debug/useful info on the http requet.
-// slower heavier but nicer source code version of DebugHandler
-func DebugHandlerTemplate(w http.ResponseWriter, r *http.Request) {
-	log.LogVf("%v %v %v %v", r.Method, r.URL, r.Proto, r.RemoteAddr)
-	hostname, _ := os.Hostname()
-	data, err := ioutil.ReadAll(r.Body)
-	if err != nil {
-		log.Errf("Error reading %v", err)
-		http.Error(w, err.Error(), http.StatusInternalServerError)
-		return
-	}
-	// Note: this looks nicer but is about 2x slower / less qps / more cpu and 25% bigger executable than doing the writes oneself:
-	const templ = `Φορτίο version {{.Version}} echo debug server on {{.Hostname}} - request from {{.R.RemoteAddr}}
-
-{{.R.Method}} {{.R.URL}} {{.R.Proto}}
-
-headers:
-
-{{ range $name, $vals := .R.Header }}{{range $val := $vals}}{{$name}}: {{ $val }}
-{{end}}{{end}}
-body:
-
-{{.Body}}
-{{if .DumpEnv}}
-environment:
-{{ range $idx, $e := .Env }}
-{{$e}}{{end}}
-{{end}}`
-	t := template.Must(template.New("debugOutput").Parse(templ))
-	err = t.Execute(w, &struct {
-		R        *http.Request
-		Hostname string
-		Version  string
-		Body     string
-		DumpEnv  bool
-		Env      []string
-	}{r, hostname, Version, DebugSummary(data, 512), r.FormValue("env") == "dump", os.Environ()})
-	if err != nil {
-		Critf("Template execution failed: %v", err)
-	}
-	w.Header().Set("Content-Type", "text/plain; charset=UTF-8")
-}
-*/
-
-// RoundDuration rounds to 10th of second. Only for positive durations.
-// TODO: switch to Duration.Round once switched to go 1.9
-func RoundDuration(d time.Duration) time.Duration {
-	tenthSec := int64(100 * time.Millisecond)
-	r := int64(d+50*time.Millisecond) / tenthSec
-	return time.Duration(tenthSec * r)
-}
-
-// DebugHandler returns debug/useful info to http client.
-func DebugHandler(w http.ResponseWriter, r *http.Request) {
-	log.LogVf("%v %v %v %v", r.Method, r.URL, r.Proto, r.RemoteAddr)
-	var buf bytes.Buffer
-	buf.WriteString("Φορτίο version ")
-	buf.WriteString(version.Long())
-	buf.WriteString(" echo debug server up for ")
-	buf.WriteString(fmt.Sprint(RoundDuration(time.Since(startTime))))
-	buf.WriteString(" on ")
-	hostname, _ := os.Hostname() // nolint: gas
-	buf.WriteString(hostname)
-	buf.WriteString(" - request from ")
-	buf.WriteString(r.RemoteAddr)
-	buf.WriteString("\n\n")
-	buf.WriteString(r.Method)
-	buf.WriteByte(' ')
-	buf.WriteString(r.URL.String())
-	buf.WriteByte(' ')
-	buf.WriteString(r.Proto)
-	buf.WriteString("\n\nheaders:\n\n")
-	// Host is removed from headers map and put here (!)
-	buf.WriteString("Host: ")
-	buf.WriteString(r.Host)
-
-	var keys []string
-	for k := range r.Header {
-		keys = append(keys, k)
-	}
-	sort.Strings(keys)
-	for _, name := range keys {
-		buf.WriteByte('\n')
-		buf.WriteString(name)
-		buf.WriteString(": ")
-		first := true
-		headers := r.Header[name]
-		for _, h := range headers {
-			if !first {
-				buf.WriteByte(',')
-			}
-			buf.WriteString(h)
-			first = false
-		}
-	}
-	data, err := ioutil.ReadAll(r.Body)
-	if err != nil {
-		log.Errf("Error reading %v", err)
-		http.Error(w, err.Error(), http.StatusInternalServerError)
-		return
-	}
-	buf.WriteString("\n\nbody:\n\n")
-	buf.WriteString(DebugSummary(data, 512))
-	buf.WriteByte('\n')
-	if r.FormValue("env") == "dump" {
-		buf.WriteString("\nenvironment:\n\n")
-		for _, v := range os.Environ() {
-			buf.WriteString(v)
-			buf.WriteByte('\n')
-		}
-	}
-	w.Header().Set("Content-Type", "text/plain; charset=UTF-8")
-	if _, err = w.Write(buf.Bytes()); err != nil {
-		log.Errf("Error writing response %v to %v", err, r.RemoteAddr)
-	}
-}
-
-// Serve starts a debug / echo http server on the given port.
-// Returns the addr where the listening socket is bound.
-// The .Port can be retrieved from it when requesting the 0 port as
-// input for dynamic http server.
-func Serve(port, debugPath string) *net.TCPAddr {
-	startTime = time.Now()
-	nPort := fnet.NormalizePort(port)
-	listener, err := net.Listen("tcp", nPort)
-	if err != nil {
-		log.Fatalf("Error occurred while listening %v: %v", nPort, err)
-	}
-	addr := listener.Addr().(*net.TCPAddr)
-	fmt.Printf("Fortio %s echo server listening on %s\n", version.Short(), addr.String())
-	go func() {
-		if debugPath != "" {
-			http.HandleFunc(debugPath, DebugHandler)
-		}
-		http.HandleFunc("/", EchoHandler)
-		if err := http.Serve(listener, nil); err != nil {
-			fmt.Println("Error starting server", err)
-		}
-	}()
-	return addr
 }
