+++ istio.io/fortio/fortio_main.go
@@ -26,6 +26,8 @@ import (
 	"runtime"
 	"strings"
 
+	"istio.io/fortio/fnet"
+
 	"istio.io/fortio/fgrpc"
 	"istio.io/fortio/fhttp"
 	"istio.io/fortio/log"
@@ -35,21 +37,31 @@ import (
 	"istio.io/fortio/version"
 )
 
-var httpOpts fhttp.HTTPOptions
-
 // -- Support for multiple instances of -H flag on cmd line:
-type flagList struct {
+type headersFlagList struct {
 }
 
-// Unclear when/why this is called and necessary
-func (f *flagList) String() string {
+func (f *headersFlagList) String() string {
 	return ""
 }
-func (f *flagList) Set(value string) error {
+func (f *headersFlagList) Set(value string) error {
 	return httpOpts.AddAndValidateExtraHeader(value)
 }
 
 // -- end of functions for -H support
+// -- Support for multiple proxies (-P) flags on cmd line:
+type proxiesFlagList struct {
+}
+
+func (f *proxiesFlagList) String() string {
+	return ""
+}
+func (f *proxiesFlagList) Set(value string) error {
+	proxies = append(proxies, value)
+	return nil
+}
+
+// -- end of functions for -P support
 
 // Prints usage
 func usage(msgs ...interface{}) {
@@ -58,7 +70,7 @@ func usage(msgs ...interface{}) {
 		version.Short(),
 		os.Args[0],
 		"where command is one of: load (load testing), server (starts grpc ping and",
-		"http echo/ui/redirect servers), grpcping (grpc client), report (report only UI",
+		"http echo/ui/redirect/proxy servers), grpcping (grpc client), report (report only UI",
 		"server), redirect (redirect only server), or curl (single URL debug).",
 		"where target is a url (http load tests) or host:port (grpc health test)",
 		"and flags are:")
@@ -81,7 +93,7 @@ var (
 	profileFlag        = flag.String("profile", "", "write .cpu and .mem profiles to file")
 	keepAliveFlag      = flag.Bool("keepalive", true, "Keep connection alive (only for fast http 1.1)")
 	halfCloseFlag      = flag.Bool("halfclose", false, "When not keepalive, whether to half close the connection (only for fast http)")
-	httpReqTimeoutFlag = flag.Duration("httpreqtimeout", fhttp.HTTPReqTimeOutDefaultValue, "Http request timeout value")
+	httpReqTimeoutFlag = flag.Duration("timeout", fhttp.HTTPReqTimeOutDefaultValue, "Connection and read timeout value (for http)")
 	stdClientFlag      = flag.Bool("stdclient", false, "Use the slower net/http standard client (works for TLS)")
 	http10Flag         = flag.Bool("http1.0", false, "Use http1.0 (instead of http 1.1)")
 	grpcFlag           = flag.Bool("grpc", false, "Use GRPC (health check) for load testing")
@@ -98,11 +110,16 @@ var (
 	curlFlag   = flag.Bool("curl", false, "Just fetch the content once")
 	labelsFlag = flag.String("labels", "",
 		"Additional config data/labels to add to the resulting JSON, defaults to target URL and hostname")
-	staticDirFlag  = flag.String("static-dir", "", "Absolute path to the dir containing the static files dir")
-	dataDirFlag    = flag.String("data-dir", defaultDataDir, "Directory where JSON results are stored/read")
-	headersFlags   flagList
+	staticDirFlag = flag.String("static-dir", "", "Absolute path to the dir containing the static files dir")
+	dataDirFlag   = flag.String("data-dir", defaultDataDir, "Directory where JSON results are stored/read")
+	headersFlags  headersFlagList
+	httpOpts      fhttp.HTTPOptions
+	proxiesFlags  proxiesFlagList
+	proxies       = make([]string, 0)
+
 	defaultDataDir = "."
 
+	followRedirectsFlag    = flag.Bool("L", false, "Follow redirects (implies -std-client) - do not use for load test")
 	allowInitialErrorsFlag = flag.Bool("allow-initial-errors", false, "Allow and don't abort on initial warmup errors")
 	autoSaveFlag           = flag.Bool("a", false, "Automatically save JSON result with filename based on labels & timestamp")
 	redirectFlag           = flag.String("redirect-port", "8081", "Redirect all incoming traffic to https URL"+
@@ -114,10 +131,20 @@ var (
 	syncFlag    = flag.String("sync", "", "index.tsv or s3/gcs bucket xml URL to fetch at startup for server modes.")
 	baseURLFlag = flag.String("base-url", "",
 		"base URL used as prefix for data/index.tsv generation. (when empty, the url from the first request is used)")
+	newMaxPayloadSizeKb = flag.Int("maxpayloadsizekb", fhttp.MaxPayloadSize/1024,
+		"MaxPayloadSize is the maximum size of payload to be generated by the EchoHandler size= argument. In Kbytes.")
+
+	// GRPC related flags
+	// To get most debugging/tracing:
+	// GODEBUG="http2debug=2" GRPC_GO_LOG_VERBOSITY_LEVEL=99 GRPC_GO_LOG_SEVERITY_LEVEL=info grpcping -loglevel debug
+	doHealthFlag  = flag.Bool("health", false, "grpc ping client mode: use health instead of ping")
+	healthSvcFlag = flag.String("healthservice", "", "which service string to pass to health check")
+	payloadFlag   = flag.String("payload", "", "Payload string to send along")
 )
 
 func main() {
 	flag.Var(&headersFlags, "H", "Additional Header(s)")
+	flag.Var(&proxiesFlags, "P", "Proxies to run, e.g -P \"localport1 dest_host1:dest_port1\" -P \"[::1]:0 www.google.com:443\" ...")
 	flag.IntVar(&fhttp.BufferSizeKb, "httpbufferkb", fhttp.BufferSizeKb,
 		"Size of the buffer (max data size) for the optimized http client in kbytes")
 	flag.BoolVar(&fhttp.CheckConnectionClosedHeader, "httpccch", fhttp.CheckConnectionClosedHeader,
@@ -138,6 +165,7 @@ func main() {
 	command := os.Args[1]
 	os.Args = append([]string{os.Args[0]}, os.Args[2:]...)
 	flag.Parse()
+	fhttp.ChangeMaxPayloadSize(*newMaxPayloadSizeKb * 1024)
 	if *quietFlag {
 		log.SetLogLevelQuiet(log.Error)
 	}
@@ -152,29 +180,48 @@ func main() {
 			os.Exit(1)
 		}
 	}
-
+	isServer := false
 	switch command {
 	case "curl":
 		fortioLoad(true, nil)
 	case "load":
 		fortioLoad(*curlFlag, percList)
 	case "redirect":
-		ui.RedirectToHTTPS(*redirectFlag)
+		isServer = true
+		fhttp.RedirectToHTTPS(*redirectFlag)
 	case "report":
+		isServer = true
 		if *redirectFlag != "disabled" {
-			go ui.RedirectToHTTPS(*redirectFlag)
+			fhttp.RedirectToHTTPS(*redirectFlag)
 		}
-		ui.Report(baseURL, *echoPortFlag, *staticDirFlag, *dataDirFlag)
+		if !ui.Report(baseURL, *echoPortFlag, *staticDirFlag, *dataDirFlag) {
+			os.Exit(1) // error already logged
+		}
 	case "server":
+		isServer = true
+		fgrpc.PingServer(*grpcPortFlag, fgrpc.DefaultHealthServiceName)
 		if *redirectFlag != "disabled" {
-			go ui.RedirectToHTTPS(*redirectFlag)
+			fhttp.RedirectToHTTPS(*redirectFlag)
 		}
-		go ui.Serve(baseURL, *echoPortFlag, *echoDbgPathFlag, *uiPathFlag, *staticDirFlag, *dataDirFlag)
-		pingServer(*grpcPortFlag)
+		if !ui.Serve(baseURL, *echoPortFlag, *echoDbgPathFlag, *uiPathFlag, *staticDirFlag, *dataDirFlag, percList) {
+			os.Exit(1) // error already logged
+		}
+		for _, proxy := range proxies {
+			s := strings.SplitN(proxy, " ", 2)
+			if len(s) != 2 {
+				log.Errf("Invalid syntax for proxy \"%s\", should be \"localAddr destHost:destPort\"", proxy)
+			}
+			fnet.ProxyToDestination(s[0], s[1])
+		}
 	case "grpcping":
 		grpcClient()
 	default:
 		usage("Error: unknown command ", command)
+	}
+	if isServer {
+		// To get a start time log/timestamp in the logs
+		log.Infof("All fortio %s servers started!", version.Long())
+		select {}
 	}
 }
 
@@ -207,6 +254,10 @@ func fortioLoad(justCurl bool, percList []float64) {
 	httpOpts.Compression = *compressionFlag
 	httpOpts.HTTPReqTimeOut = *httpReqTimeoutFlag
 	httpOpts.Insecure = *httpsInsecureFlag
+	if *followRedirectsFlag {
+		httpOpts.FollowRedirects = true
+		httpOpts.DisableFastClient = true
+	}
 	if justCurl {
 		fetchURL(&httpOpts)
 		return
@@ -257,10 +308,11 @@ func fortioLoad(justCurl bool, percList []float64) {
 	var err error
 	if *grpcFlag {
 		o := fgrpc.GRPCRunnerOptions{
-			RunnerOptions: ro,
-			Destination:   url,
-			Secure:        *grpcSecureFlag,
-			Service:       *healthSvcFlag,
+			RunnerOptions:      ro,
+			Destination:        url,
+			Secure:             *grpcSecureFlag,
+			Service:            *healthSvcFlag,
+			AllowInitialErrors: *allowInitialErrorsFlag,
 		}
 		res, err = fgrpc.RunGRPCTest(&o)
 	} else {
@@ -317,5 +369,27 @@ func fortioLoad(justCurl bool, percList []float64) {
 			}
 		}
 		fmt.Fprintf(out, "Successfully wrote %d bytes of Json data to %s\n", n, jsonFileName)
+	}
+}
+
+func grpcClient() {
+	if len(flag.Args()) != 1 {
+		usage("Error: fortio grpcping needs host argument in the form of host, host:port or ip:port")
+	}
+	host := flag.Arg(0)
+	count := int(*exactlyFlag)
+	if count <= 0 {
+		count = 1
+	}
+	tls := *grpcSecureFlag
+	var err error
+	if *doHealthFlag {
+		_, err = fgrpc.GrpcHealthCheck(host, tls, *healthSvcFlag, count)
+	} else {
+		_, err = fgrpc.PingClientCall(host, tls, count, *payloadFlag)
+	}
+	if err != nil {
+		// already logged
+		os.Exit(1)
 	}
 }
