+++ istio.io/fortio/ui/uihandler.go
@@ -40,7 +40,6 @@ import (
 
 	"istio.io/fortio/fgrpc"
 	"istio.io/fortio/fhttp"
-	"istio.io/fortio/fnet"
 	"istio.io/fortio/log"
 	"istio.io/fortio/periodic"
 	"istio.io/fortio/stats"
@@ -77,6 +76,8 @@ var (
 	runs           = make(map[int64]*periodic.RunnerOptions)
 	// Base URL used for index - useful when running under an ingress with prefix
 	baseURL string
+
+	defaultPercentileList []float64
 )
 
 const (
@@ -104,20 +105,6 @@ func getResourcesDir(override string) string {
 	return ""
 }
 
-// HTMLEscapeWriter is an io.Writer escaping the output for safe html inclusion.
-type HTMLEscapeWriter struct {
-	NextWriter io.Writer
-	Flusher    http.Flusher
-}
-
-func (w *HTMLEscapeWriter) Write(p []byte) (int, error) {
-	template.HTMLEscape(w.NextWriter, p)
-	if w.Flusher != nil {
-		w.Flusher.Flush()
-	}
-	return len(p), nil
-}
-
 // TODO: auto map from (Http)RunnerOptions to form generation and/or accept
 // JSON serialized options as input.
 
@@ -137,7 +124,7 @@ const (
 
 // Handler is the main UI handler creating the web forms and processing them.
 func Handler(w http.ResponseWriter, r *http.Request) {
-	LogRequest(r, "UI")
+	fhttp.LogRequest(r, "UI")
 	mode := menu
 	JSONOnly := false
 	DoSave := (r.FormValue("save") == "on")
@@ -183,8 +170,11 @@ func Handler(w http.ResponseWriter, r *http.Request) {
 		log.Fatalf("expected http.ResponseWriter to be an http.Flusher")
 	}
 	out := io.Writer(os.Stderr)
+	if len(percList) == 0 && !strings.Contains(r.URL.RawQuery, "p=") {
+		percList = defaultPercentileList
+	}
 	if !JSONOnly {
-		out = io.Writer(&HTMLEscapeWriter{NextWriter: w, Flusher: flusher})
+		out = fhttp.NewHTMLEscapeWriter(w)
 	}
 	n, _ := strconv.ParseInt(r.FormValue("n"), 10, 64) // nolint: gas
 	if strings.TrimSpace(url) == "" {
@@ -290,7 +280,7 @@ func Handler(w http.ResponseWriter, r *http.Request) {
 				log.Errf("Error adding custom headers: %v", err)
 			}
 		}
-		onBehalfOf(httpopts, r)
+		fhttp.OnBehalfOf(httpopts, r)
 		if !JSONOnly {
 			flusher.Flush()
 		}
@@ -402,7 +392,7 @@ func DataList() (dataList []string) {
 
 // BrowseHandler handles listing and rendering the JSON results.
 func BrowseHandler(w http.ResponseWriter, r *http.Request) {
-	LogRequest(r, "Browse")
+	fhttp.LogRequest(r, "Browse")
 	path := r.URL.Path
 	if (path != uiPath) && (path != (uiPath + "browse")) {
 		if strings.HasPrefix(path, "/fortio") {
@@ -438,23 +428,10 @@ func BrowseHandler(w http.ResponseWriter, r *http.Request) {
 	}
 }
 
-// LogRequest logs the incoming request, including headers when loglevel is verbose
-func LogRequest(r *http.Request, msg string) {
-	log.Infof("%s: %v %v %v %v (%s)", msg, r.Method, r.URL, r.Proto, r.RemoteAddr,
-		r.Header.Get("X-Forwarded-Proto"))
-	if log.LogVerbose() {
-		for name, headers := range r.Header {
-			for _, h := range headers {
-				log.LogVf("Header %v: %v\n", name, h)
-			}
-		}
-	}
-}
-
 // LogAndAddCacheControl logs the request and wrapps an HTTP handler to add a Cache-Control header for static files.
 func LogAndAddCacheControl(h http.Handler) http.Handler {
 	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		LogRequest(r, "Static")
+		fhttp.LogRequest(r, "Static")
 		path := r.URL.Path
 		if path == faviconPath {
 			r.URL.Path = "/static/img" + faviconPath // fortio/version expected to be stripped already
@@ -543,7 +520,7 @@ func sendTSVDataIndex(urlPrefix string, w http.ResponseWriter) {
 // LogAndFilterDataRequest logs the data request.
 func LogAndFilterDataRequest(h http.Handler) http.Handler {
 	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		LogRequest(r, "Data")
+		fhttp.LogRequest(r, "Data")
 		path := r.URL.Path
 		if strings.HasSuffix(path, "/") || strings.HasSuffix(path, "/index.html") {
 			sendHTMLDataIndex(w)
@@ -575,41 +552,6 @@ func LogAndFilterDataRequest(h http.Handler) http.Handler {
 		}
 		h.ServeHTTP(w, r)
 	})
-}
-
-// FetcherHandler is the handler for the fetcher/proxy.
-func FetcherHandler(w http.ResponseWriter, r *http.Request) {
-	LogRequest(r, "Fetch")
-	hj, ok := w.(http.Hijacker)
-	if !ok {
-		log.Critf("hijacking not supported")
-		return
-	}
-	conn, _, err := hj.Hijack()
-	if err != nil {
-		log.Errf("hijacking error %v", err)
-		return
-	}
-	// Don't forget to close the connection:
-	defer conn.Close() // nolint: errcheck
-	url := r.URL.String()[len(fetchPath):]
-	opts := fhttp.NewHTTPOptions("http://" + url)
-	opts.HTTPReqTimeOut = 5 * time.Minute
-	onBehalfOf(opts, r)
-	client := fhttp.NewClient(opts)
-	if client == nil {
-		return // error logged already
-	}
-	_, data, _ := client.Fetch()
-	_, err = conn.Write(data)
-	if err != nil {
-		log.Errf("Error writing fetched data to %v: %v", r.RemoteAddr, err)
-	}
-	client.Close()
-}
-
-func onBehalfOf(o *fhttp.HTTPOptions, r *http.Request) {
-	_ = o.AddAndValidateExtraHeader("X-On-Behalf-Of: " + r.RemoteAddr) // nolint: gas
 }
 
 // TODO: move tsv/xml sync handling to their own file (and possibly package)
@@ -653,7 +595,7 @@ func Sync(out io.Writer, u string, datadir string) bool {
 
 // SyncHandler handles syncing/downloading from tsv url.
 func SyncHandler(w http.ResponseWriter, r *http.Request) {
-	LogRequest(r, "Sync")
+	fhttp.LogRequest(r, "Sync")
 	flusher, ok := w.(http.Flusher)
 	if !ok {
 		log.Fatalf("expected http.ResponseWriter to be an http.Flusher")
@@ -672,7 +614,7 @@ func SyncHandler(w http.ResponseWriter, r *http.Request) {
 	w.Write([]byte("Fetch of index/bucket url ... ")) // nolint: gas, errcheck
 	flusher.Flush()
 	o := fhttp.NewHTTPOptions(uStr)
-	onBehalfOf(o, r)
+	fhttp.OnBehalfOf(o, r)
 	// If we had hundreds of thousands of entry we should stream, parallelize (connection pool)
 	// and not do multiple passes over the same data, but for small tsv this is fine.
 	// use std client to change the url and handle https:
@@ -848,15 +790,18 @@ func downloadOne(w http.ResponseWriter, client *fhttp.Client, name string, u str
 
 // Serve starts the fhttp.Serve() plus the UI server on the given port
 // and paths (empty disables the feature). uiPath should end with /
-// (be a 'directory' path)
-func Serve(baseurl, port, debugpath, uipath, staticRsrcDir string, datadir string) {
+// (be a 'directory' path). Returns true if server is started successfully.
+func Serve(baseurl, port, debugpath, uipath, staticRsrcDir string, datadir string, percentileList []float64) bool {
 	baseURL = baseurl
 	startTime = time.Now()
-	hostPort := fnet.NormalizePort(port)
-	if uipath == "" {
-		fhttp.Serve(hostPort, debugpath)
-		return
+	mux, addr := fhttp.Serve(port, debugpath)
+	if addr == nil {
+		return false // Error already logged
 	}
+	if uipath == "" {
+		return true
+	}
+	fhttp.SetupPPROF(mux)
 	uiPath = uipath
 	dataDir = datadir
 	if uiPath[len(uiPath)-1] != '/' {
@@ -864,9 +809,9 @@ func Serve(baseurl, port, debugpath, uipath, staticRsrcDir string, datadir strin
 		uiPath += "/"
 	}
 	debugPath = ".." + debugpath // TODO: calculate actual path if not same number of directories
-	http.HandleFunc(uiPath, Handler)
+	mux.HandleFunc(uiPath, Handler)
 	fetchPath = uiPath + fetchURI
-	http.HandleFunc(fetchPath, FetcherHandler)
+	mux.Handle(fetchPath, http.StripPrefix(fetchPath, http.HandlerFunc(fhttp.FetcherHandler)))
 	fhttp.CheckConnectionClosedHeader = true // needed for proxy to avoid errors
 
 	logoPath = version.Short() + "/static/img/logo.svg"
@@ -881,8 +826,8 @@ func Serve(baseurl, port, debugpath, uipath, staticRsrcDir string, datadir strin
 	if staticRsrcDir != "" {
 		fs := http.FileServer(http.Dir(staticRsrcDir))
 		prefix := uiPath + version.Short()
-		http.Handle(prefix+"/static/", LogAndAddCacheControl(http.StripPrefix(prefix, fs)))
-		http.Handle(faviconPath, LogAndAddCacheControl(fs))
+		mux.Handle(prefix+"/static/", LogAndAddCacheControl(http.StripPrefix(prefix, fs)))
+		mux.Handle(faviconPath, LogAndAddCacheControl(fs))
 		var err error
 		mainTemplate, err = template.ParseFiles(path.Join(staticRsrcDir, "templates/main.html"))
 		if err != nil {
@@ -892,38 +837,42 @@ func Serve(baseurl, port, debugpath, uipath, staticRsrcDir string, datadir strin
 		if err != nil {
 			log.Critf("Unable to parse browse template: %v", err)
 		} else {
-			http.HandleFunc(uiPath+"browse", BrowseHandler)
+			mux.HandleFunc(uiPath+"browse", BrowseHandler)
 		}
 		syncTemplate, err = template.ParseFiles(path.Join(staticRsrcDir, "templates/sync.html"))
 		if err != nil {
 			log.Critf("Unable to parse sync template: %v", err)
 		} else {
-			http.HandleFunc(uiPath+"sync", SyncHandler)
+			mux.HandleFunc(uiPath+"sync", SyncHandler)
 		}
 	}
 	if dataDir != "" {
 		fs := http.FileServer(http.Dir(dataDir))
-		http.Handle(uiPath+"data/", LogAndFilterDataRequest(http.StripPrefix(uiPath+"data", fs)))
+		mux.Handle(uiPath+"data/", LogAndFilterDataRequest(http.StripPrefix(uiPath+"data", fs)))
 	}
-	addr := fhttp.Serve(hostPort, debugpath)
-	setHostAndPort(hostPort, addr)
-	uiMsg := fmt.Sprintf("UI starting - visit:\nhttp://%s%s", urlHostPort, uiPath)
+	setHostAndPort(port, addr)
+	uiMsg := fmt.Sprintf("UI started - visit:\nhttp://%s%s", urlHostPort, uiPath)
 	if !strings.Contains(port, ":") {
 		uiMsg += "   (or any host/ip reachable on this server)"
 	}
 	fmt.Printf(uiMsg + "\n")
-
+	defaultPercentileList = percentileList
+	return true
 }
 
 // Report starts the browsing only UI server on the given port.
 // Similar to Serve with only the read only part.
-func Report(baseurl, port, staticRsrcDir string, datadir string) {
-	http.DefaultServeMux = http.NewServeMux() // drop the pprof default handlers
+func Report(baseurl, port, staticRsrcDir string, datadir string) bool {
+	// drop the pprof default handlers [shouldn't be needed with custom mux but better safe than sorry]
+	http.DefaultServeMux = http.NewServeMux()
 	baseURL = baseurl
 	extraBrowseLabel = ", report only limited UI"
-	hostPort := fnet.NormalizePort(port)
-	setHostAndPort(hostPort, nil)
-	uiMsg := fmt.Sprintf("Browse only UI starting - visit:\nhttp://%s/", urlHostPort)
+	mux, addr := fhttp.HTTPServer("report", port)
+	if addr == nil {
+		return false
+	}
+	setHostAndPort(port, addr)
+	uiMsg := fmt.Sprintf("Browse only UI started - visit:\nhttp://%s/", urlHostPort)
 	if !strings.Contains(port, ":") {
 		uiMsg += "   (or any host/ip reachable on this server)"
 	}
@@ -935,55 +884,31 @@ func Report(baseurl, port, staticRsrcDir string, datadir string) {
 	staticRsrcDir = getResourcesDir(staticRsrcDir)
 	fs := http.FileServer(http.Dir(staticRsrcDir))
 	prefix := uiPath + version.Short()
-	http.Handle(prefix+"/static/", LogAndAddCacheControl(http.StripPrefix(prefix, fs)))
-	http.Handle(faviconPath, LogAndAddCacheControl(fs))
+	mux.Handle(prefix+"/static/", LogAndAddCacheControl(http.StripPrefix(prefix, fs)))
+	mux.Handle(faviconPath, LogAndAddCacheControl(fs))
 	var err error
 	browseTemplate, err = template.ParseFiles(path.Join(staticRsrcDir, "templates/browse.html"))
 	if err != nil {
 		log.Critf("Unable to parse browse template: %v", err)
 	} else {
-		http.HandleFunc(uiPath, BrowseHandler)
+		mux.HandleFunc(uiPath, BrowseHandler)
 	}
 	fsd := http.FileServer(http.Dir(dataDir))
-	http.Handle(uiPath+"data/", LogAndFilterDataRequest(http.StripPrefix(uiPath+"data", fsd)))
-	if err := http.ListenAndServe(hostPort, nil); err != nil {
-		log.Critf("Error starting report server: %v", err)
-	}
-}
-
-// -- Redirection to https feature --
-
-// RedirectToHTTPSHandler handler sends a redirect to same URL with https.
-func RedirectToHTTPSHandler(w http.ResponseWriter, r *http.Request) {
-	dest := "https://" + r.Host + r.URL.String()
-	LogRequest(r, "Redirecting to "+dest)
-	http.Redirect(w, r, dest, http.StatusSeeOther)
-}
-
-// RedirectToHTTPS Sets up a redirector to https on the given port.
-// (Do not create a loop, make sure this is addressed from an ingress)
-func RedirectToHTTPS(port string) {
-	m := http.NewServeMux()
-	m.HandleFunc("/", RedirectToHTTPSHandler)
-	s := &http.Server{
-		Addr:    fnet.NormalizePort(port),
-		Handler: m,
-	}
-	fmt.Printf("Https redirector running on %v\n", s.Addr)
-	if err := s.ListenAndServe(); err != nil {
-		log.Critf("Error starting redirect server: %v", err)
-	}
-	fmt.Printf("Not reached, https redirector exiting - was on %v\n", s.Addr)
+	mux.Handle(uiPath+"data/", LogAndFilterDataRequest(http.StripPrefix(uiPath+"data", fsd)))
+	return true
 }
 
 // setHostAndPort takes hostport in the form of hostname:port, ip:port or :port,
-// sets the urlHostPort variable and returns hostport unmodified.
+// sets the urlHostPort variable.
 func setHostAndPort(inputPort string, addr *net.TCPAddr) {
 	urlHostPort = inputPort
 	portStr := inputPort
 	if addr != nil {
 		urlHostPort = addr.String()
 		portStr = fmt.Sprintf(":%d", addr.Port)
+	}
+	if !strings.Contains(inputPort, ":") {
+		inputPort = ":" + inputPort
 	}
 	if strings.HasPrefix(inputPort, ":") {
 		urlHostPort = "localhost" + portStr
